requirements.txt
click
requests
python-decouple